split SD commands into common and SD-specific modules
