+++ src/common_cmd.rs
@@ -21,10 +21,6 @@ pub struct R1;
 pub struct R2;
 /// R3: OCR register
 pub struct R3;
-/// R6: Published RCA response
-pub struct R6;
-/// R7: Card interface condition
-pub struct R7;
 
 pub trait Resp {
     const LENGTH: ResponseLen = ResponseLen::R48;
@@ -40,8 +36,6 @@ impl Resp for R2 {
 
 impl Resp for R1 {}
 impl Resp for R3 {}
-impl Resp for R6 {}
-impl Resp for R7 {}
 
 /// Command Response type
 #[derive(Eq, PartialEq, Copy, Clone)]
@@ -72,25 +66,9 @@ pub fn all_send_cid() -> Cmd<R2> {
     cmd(2, 0)
 }
 
-/// CMD3: Send RCA
-pub fn send_relative_address() -> Cmd<R6> {
-    cmd(3, 0)
-}
-
-/// CMD6: Switch Function Command
-pub fn cmd6(arg: u32) -> Cmd<R1> {
-    cmd(6, arg)
-}
-
 /// CMD7: Select or deselect card
 pub fn select_card(rca: u16) -> Cmd<R1> {
     cmd(7, u32::from(rca) << 16)
-}
-
-/// CMD8: Sends memory card interface conditions
-pub fn send_if_cond(voltage: u8, checkpattern: u8) -> Cmd<R7> {
-    let arg = u32::from(voltage & 0xF) << 8 | u32::from(checkpattern);
-    cmd(8, arg)
 }
 
 /// CMD9: Send CSD
@@ -101,11 +79,6 @@ pub fn send_csd(rca: u16) -> Cmd<R2> {
 /// CMD10: Send CID
 pub fn send_cid(rca: u16) -> Cmd<R2> {
     cmd(10, u32::from(rca) << 16)
-}
-
-/// CMD11: Switch to 1.8V bus signaling level
-pub fn voltage_switch() -> Cmd<R1> {
-    cmd(11, 0)
 }
 
 /// CMD12: Stop transmission
@@ -139,26 +112,6 @@ pub fn read_multiple_blocks(addr: u32) -> Cmd<R1> {
     cmd(18, addr)
 }
 
-/// CMD19: Send tuning pattern
-pub fn send_tuning_block(addr: u32) -> Cmd<R1> {
-    cmd(19, addr)
-}
-
-/// CMD20: Speed class control
-pub fn speed_class_control(arg: u32) -> Cmd<R1> {
-    cmd(20, arg)
-}
-
-/// CMD22: Address extension
-pub fn address_extension(arg: u32) -> Cmd<R1> {
-    cmd(22, arg)
-}
-
-/// CMD23: Address extension
-pub fn set_block_count(blockcount: u32) -> Cmd<R1> {
-    cmd(23, blockcount)
-}
-
 /// CMD24: Write block
 pub fn write_single_block(addr: u32) -> Cmd<R1> {
     cmd(24, addr)
@@ -177,35 +130,4 @@ pub fn program_csd() -> Cmd<R1> {
 /// CMD55: App Command. Indicates that next command will be a app command
 pub fn app_cmd(rca: u16) -> Cmd<R1> {
     cmd(55, u32::from(rca) << 16)
-}
-
-/// ACMD6: Bus Width
-/// * `bw4bit` - Enable 4 bit bus width
-pub fn set_bus_width(bw4bit: bool) -> Cmd<R1> {
-    let arg = if bw4bit { 0b10 } else { 0b00 };
-    cmd(6, arg)
-}
-
-/// ACMD13: SD Status
-pub fn sd_status() -> Cmd<R1> {
-    cmd(13, 0)
-}
-
-/// ACMD41: App Op Command
-///
-/// * `high_capacity` - Host supports high capacity cards
-/// * `xpc` - Controls the maximum power and default speed mode of SDXC and SDUC cards.
-/// * `s18r` - Switch to 1.8V signaling
-/// * `voltage_window` - The voltage window the host supports.
-pub fn sd_send_op_cond(high_capacity: bool, xpc: bool, s18r: bool, voltage_window: u32) -> Cmd<R3> {
-    let arg = u32::from(high_capacity) << 30
-        | u32::from(xpc) << 28
-        | u32::from(s18r) << 24
-        | voltage_window & 0x00FF_FFFF;
-    cmd(41, arg)
-}
-
-/// ACMD51: Reads the SCR
-pub fn send_scr() -> Cmd<R1> {
-    cmd(51, 0)
 }
